split out classes for redis async queue
